testando o jPA

+++ LIMPO.sql
@@ -25,7 +25,7 @@ CREATE TABLE curso (
 CREATE TABLE matricula (
   idcurso INT(11) NOT NULL,
   idaluno INT(11) NOT NULL,
-  nota INT NULL,
+  nota DOUBLE NULL,
   PRIMARY KEY (idcurso, idaluno),
   CONSTRAINT FK_IDALUNO
     FOREIGN KEY (idaluno)
